New artifact: The Ring of P'hul This is another powerful magic item from 'The Lords of Dus' series of books written by Lawrence Watt-Evans. In the series, the ring was extremely powerful, and spread death and disease over great distances, and was also very hard to control. Those wielding the ring could wind up wiping out themselves as well as the enemy they were targeting. My take on it is in the other direction - this ring protects the wearer from poison and disease, and can also regenerate hit points. Making this the Barbarian quest artifact, as the default one has little value to most players.
Current issue - because this artifact has a common object name in its description, trying to wish for it generates a random regular ring instead. At this time, I don't know how to resolve that, but this will be addressed. Otherwise, this new artifact behaves as intended and functions as the Barbarian quest artifact.