Nazgul refresh Some changes to the Nazgul, as well as some other additions: * the Nazgul now has a scream attack that can stun you (taken from SporkHack) * the Nazgul has been buffed up a little bit, and can see invisible * the Nazgul now spawn riding a steed, a new monster called a 'Fell Beast' * neither the Nazgul nor its mount can be genocided
In SporkHack (and UnNetHack) the Nazgul's scream was the bane of many players, and it quickly made its way towards the top of a players genocide list. If you could see the Nazgul, by any means, its scream would affect you. I spent a fair amount of time cleaning up this segment of code, setting the conditions so that the Nazgul would only scream if it could see you (not vice versa), AND it would only affect you if you weren't deaf or stun resistant. The frequency of its scream has been toned down as well.

The fell beasts that the Nazgul ride comes straight from J.R.R Tolkien lore - they're not dragons, but they do look like them somewhat. They have a bite (beak) and claw attack and are resistant to various forms of attack. A worthy steed for its Nazgul masters.