Dragon scale mail features Another pull from SporkHack, but modified. Certain DSM can exhibit secondary extrinsics that are beneficial to the player that wears it. Secondary extrinsics are as follows (* denotes same as SporkHack, + denotes custom/new): * Blue       - speed * * Black      - passive disintegration attack (1d4) + * Gray       - (no change) * Green      - passive poison attack (1d4) * and + * Orange     - passive slow attack * * Red        - +1d6 to all melee attack * * Silver     - (no change) * White      - freezes water/lava underfoot * * Yellow     - stone resistance + * Gold       - (no change [radius 2 light]) * * Shimmering - (no change)
The passive poison attack has the typical 1 in 20 chance of insta-killing the attacker if they are not poison resistant. The passive disintegration attack has a 1 in 40 chance of outright disintegrating the attacker if they are not disintegration resistant. Take note, currently none of the secondary extrinsics take affect when the DSM is worn by monsters or other NPC's.

One thing I've added in conjunction to this that SporkHack did not do, was to give certain dragons the same benefits of
the secondary extrinsics of their scales. Here's the breakdown:
* Blue    - speed adjustment from 9 to 14
* Black   - passive corrosive damage to non-fooproof weapons
* Green   - passive non-instadeath poison attack (1d6)
* Orange  - passive slow attack
* Yellow  - passive corrosive damage to non-fooproof weapons

Other bits: tweaked AD_SLOW attack type so it behaves as intended. May need further adjustment.