New intrinsic: Food Sense. I created a permanent intrinsic that acts the same way as casting the spell 'Detect Food' while skilled or expert in divination, or if you had read a blessed scroll of detect food. If you're about to eat a corpse and its dangerous in any way, you'll get a warning about it, with a yes/no option to continue. The difference here is that this intrinsic (Food_sense) is permanent and does not wear off after one use.
For the time being I added this to the 'Giant' player race for testing purposes only. I have another new player race in mind that I am going to give this intrinsic to.

Known issues: if you detect a corpse as dangerous and you have the option to eat it anyways (y/n?), regardless of how you answer, you eat it. That's a problem. Will figure this out.