Crystal Plate Mail armor. In some past commits, I've given special consideration to crystal plate mail armor to make it a bit more desireable to wear (immune to the revamped monster 'destroy armor' spell, no spell percentage chance penalty when wearing it, etc). Decided to go all-in with it, and make it something the player would actually want to seek out and use.
Back story: crystal plate mail is forged from dilithium crystal, and can only be made by the most skilled of blacksmiths in conjunction with a powerful wizard. Dilithium is virtually transparent to magic; it almost acts like it doesn't exist. To some, this may seem like a bad thing, but for others it's a blessing.

Changes in this commit:
* Material changed from GLASS to GEMSTONE; no longer shatters like other glass objects.
* Weight dropped to 150 aum. <--- original weight was #1 reason why no one would wear it
* Immune to scrolls of destroy armor and black dragon disintegration breath.
* Rarity increased, along with its base cost.

I almost made it immune to ALL forms of magic, but then you wouldn't be able to enchant it, or uncurse it (and since destroy armor scrolls no longer work on this, that could have become a very bad situation). These plus past changes should finally make this into something the player might want to use, especially as a spellcaster.