Magic 8-Ball. Was inspired by the !8ball function Beholder bot has in the #hardfought IRC channel. Here's the back story:
The Oracle is not as all-powerful as she would have you think. All of her knowledge comes from a magical artifact known as the Magic 8-Ball. All of the rumors The Oracle knows come from it. If only our hero could obtain all of that power and knowledge for themselves...

The Magic 8-Ball is not that powerful at all compared to other artifacts. It's a small plastic orb that weighs 20 aum. Its main function is to deliver a rumor each time the player applies it. It's subject to the same blessed/uncursed/cursed rules when giving a rumor as the MKoT and The Magic Mirror of Merlin. It conveys warning when in your inventory. Here's the special bit - if blessed, there's a 1 in 1000 chance that the Magic 8-Ball will grant the player a single wish when applied for a rumor (uncursed or cursed, no chance). Should the player get a wish from the 8-Ball, it disappears, never to be seen again.

Doesn't sound like a lot, but I've also added over 20 new true rumors in this commit that are specific to the changes that have been made to EvilHack (these will show up in fortune cookies and other rumor-giving artifacts also). My unique and silly attempt to give the player some incentive to discover new things about the new variant they're playing. The Oracle isn't going to give up it's source of power so easily; The Oracle has been beefed up a little bit so the player will need to work for it somewhat if they want the 8-Ball. Since this adds another guaranteed artifact early on, I will probably tweak the artifact wishing formula to compensate.

Other bits - wands of wishing and magic markers are a little less common than they used to be (random generation). SporkHack made random generation of these zero, not going to be that drastic. Also fixed a bug concerning shambling horrors that would randomly crash the game when one was generated (had to do with alignment value and random monster generation).