Elves hate iron; Orcs hate mithril. From UnNetHack (heavily modified) - if the player is an elf, any direct contact with iron (wielded or worn) prevents hit point regeneration. Same for orcs and mithril.
Because elven knights are a thing in EvilHack, their starting gear has been adjusted. I've also introduced an elven long sword, which basically has the same stats as a katana. Higher level elves have a chance of spawning with one instead of an elven broadsword, but overall these are more rare than other elven weapons. To obtain Dirge, it has to be a regular long sword (sorry 'bout that elven knights).