Object Materials patch. From xNetHack - this patch allows a multitude of weapons, armor and other items to generate with a material other than its given base material. Other variants include object materials (most notably GruntHack) but the work done for xNetHack and pulled into EvilHack significantly improves upon that work.
This isn't a one-for-one rip - I've tweaked and altered a few things in ways that make sense to me. Things like, you can include the object material when making a wish (if it's allowed for that object), dragonhide is a thing like it is in GruntHack (but very rare). The weights of certain materials and the frequency/type of materials that are possible for certain items also differ here from xNetHack's implementation. So on and so forth. This work also brings the version of EvilHack up to 0.2.0.

I'm sure there's more work to be done, this was one hell of a pull from xNetHack and it's possible I missed a thing or two. Also, there's a segfault that occurs now when trying to play in wizmode and using a wizkit text file (see https://termbin.com/dak6 for the coredump), and if the segfault doesn't happen, many items in inventory that have no business being leather, are leather.