Magic markers no longer randomly generated. From SporkHack (sort of) - in SporkHack, magic markers were set as never randomly occuring, and your role would never start with one either. Only way to get one was to either wish for it or (maybe) polypile for one.
For EvilHack - random generation is gone, as is certain roles spawning with one. However, in Sokoban there's a 50/50 chance of one of the three prizes at the end being a magic marker. I did not change anything in regards to polypiling behavior, so the player could potentially get one via that method.

Another evil bit - wands of nothing are now prices at 500zm just like wands of wishing and death. Have fun price-ID'ing ;)