Monsters can offensively use wands of polymorph. From GruntHack (modified) - there was already logic here in regards to offensively using potions of polymorph, was time to allow for wands as well.
As in vanilla NetHack, if a monster if of difficultly level < 6, it will use the wand on itself first in an attempt to turn into something more powerful. If the monster if of greater difficulty, and meets other criteria (has hands, isn't mindless, can see you and you're in range, etc) it will zap the wand at you instead. Just like with other offensive wands, if the monster using the wand has a scroll of charging, it will use it to recharge the wand.

Touched up some logic with a couple other things as well.