Gehennom revamp - round two. This commit adds the four special demon lord/prince levels from GruntHack - Dispater, Yeenoghu, Geryon, and Demogorgon. The levels were edited primarily to conform to 3.6.x rules, and a couple extra edits to Dispater's level so it fits in with the new 'hellfill' regular Gehennom levels. Otherwise, this is a straight port over to EvilHack.
Expect to see a 'round three', I want to tweak these, maybe change the monster placement, edit the maps some. Also in GruntHack, your demon lord/prince types would not teleport to the upstairs when they got low on hit points. I don't plan on changing vanilla behavior to act like GruntHack, so some of these levels (read: Demogorgon's lair) would be a huge tedious pain in the ass if you had to backtrack to the upstairs to kill that particular demon lord or prince. More to follow.