New mines end map: Orc Temple Created by Khor - this adds a fourth random end to the gnomish mines: an ancient temple overseen by orcs, which hides a terrible, powerful secret...
I've edited this slightly in the following ways:
* code altered to conform with 3.6.x map rules
* minor tweaks to types of traps and containers
* a few changes to some of the monsters that will spawn (more goblin types are represented)
* since the player can't use the E-word until they discover it, we'll throw them a bone here: Elbereth is etched into the altar within the hidden temple. There's a cursed scroll of 'scare monster' here also, as this is the spot where the figurine rests, and we don't need the temple shamans activating it since that's a thing now (testing was entertaining to say the least).

This map as well the the gnome king's wine cellar has a spot where Elbereth is etched into the ground. So the player has a 50/50 chance of learning it if they complete the mines.

This is a really fun map, and I'm surprised it hasn't made its way into several of the other variants out there. Enjoy.