Fix: 'placing monster over another' and 'can't place branch' impossible messages. The first issue 'placing monster over another' is caused by monsters riding other monsters as steeds. It doesn't adversely affect gameplay, but the error messages are disconcerting to say the least, and would appear everytime you left a level with a monster riding another and then coming back to it. All we're doing here is suppressing the error message from occuring with monster steeds. The DevTeam has done the same thing with vault guards, I see no harm in doing the same. There *might* be some cases of this also with zombie corpse revival, but I haven't seen any yet while testing. Can always revisit this if necessary.
The second issue 'can't place branch' was caused by one random vault room - the ice room, which had a random fire trap somewhere in the room 50% of the time. Again, the random room would be created, but if anything else generated in that room, such as stairs or the trap, the error message would occur. Easiest just to kill that one room. There are other randomly generated ice-type rooms created via different methods, so no big loss.

In this commit I also upped the try count for generating and properly placing rooms and such from 100 to 500 - there are almost 1500 different spots to place something in a regular sized terminal (80x24) when playing nethack. Trying 500 times vs 100 seems to me to make more sense given the area we're playing with. The servers can more than handle the extra tries regardless.