Fix priests and shopkleepers stepping on zombies.
This is a fix for one case of the impossible("placing monster over another at...")
errors we have been seeing.  Priests and shks will now attack zombies instead
of occupying their space. The attack will only occur if the priest/shk was going
to move to that space anyway (because having your shop or temple overrun by
legions of undead is not a bad thing per se, only if they happen to get in your
way). This could stand some rethinking, the primary goal here was to fix the bug.