Artifact properties when in offhand and two-weaponing - part three. Artifacts now convey their special effects and abilities when wielded in the offhand when twoweaponing. Current issue - effects and abilities are still present when the hero switches from twoweaponing to just their primary weapon and the offhand artifact is flagged as 'alternate weapon; not wielded'. This is giving the player a freebie (wield one weapon + shield and still enjoy benefits of the artifact flagged as alternate weapon). I've tried several different methods to try and have this work only when in twoweaponing mode, but nothing has worked yet. So far it's either all or none. Took out some code initially added from GruntHack that I now think only worked with regular items w/ object properties. More to do, but getting there though...
Also in this commit: re-worked the 'you are aware of the presense of <insert monster type here>.' when checking attributes in wizmode or via enlightenment to take into account that the player can now have up to three items active at once that can detect three different classes of monsters (artifact weapons in primary and offhand, and then wearing Dragonbane). There's logic in here now too to handle wielding Orcrist and Sting in twoweaponing mode - they'll both light up in the presence of S_ORC whether in primary or offhand (could use a bit more work).