Objects can be completely destroyed via rusting/rotting/corroding. From GruntHack (heavily modified) - any object in open inventory, whether it's worn, wielded, or just out in the open can be completely destroyed via rust/rot/corrode attacks or events if those objects are not fooproofed. This holds true for monsters as well as the player. Sources include rust traps, fountains, swimming in water, attacks from various monsters (rust monsters, water elementals, brown or black puddings, any monster with an acid-based attack), player and monster spells that can corrode or rust, and potions or other objects that can do the same.
In some ways this can be useful - if the player is wearing a cursed iron piece of armor or wielding a cursed iron weapon, or stuck with an iron ball/chain, you could dip either into a fountain or pool until they rust away completely. Overall though, players will need to be more careful concerning gear and what can affect it. This effectively makes scrolls of enchant armor/weapon, the spell 'repair armor', potions of oil and cans of grease more valuable.

Known bugs: monsters inventory will throw an error if it erodes away completely (Example: Eric the Ronin swings his katana at the rust monster.  Eric the Ronin strikes the rust monster.  Eric the Ronin's katana rusts away!  obfree: deleting worn obj (38: 256)  Program in disorder!  (Saving and reloading may fix this problem.)) - game will continue normally, no crashes. Had this problem before with the monster spell 'destroy armor' against other monsters and that was resolved. This one is a bit more complex but in the same vein, so I imagine a fix will be found soon.

Side effect of this commit: I resolved a segfault that crept in during testing, had to do with xname() and some custom code for the sokoban end prize names. Had to revert a bit of that to squash the segfault. Now, all sokoban end prizes will show as 'sokoban prize <object>' until the player actually touches it, not just direct line of sight.