Rogue new skill: Thievery. Saw this patch up on the NetHack Patch Database (https://bilious.alt.org/?37) which was submitted for 3.4.1 way back in early 2005. The comments were positive, but as far as I know, it was never realized in any variant. Figured I'd incorporate it here - if you're a Rogue, you should be able to steal. Here's the details:
* Rogue role trades out bare handed attack skill for thievery.
* Thievery can be trained up to Master level.
* To use the skill, a Rogue will 'attack' its target bare handed (to physically attack bare handed to cause damage, use shift+F).
* The higher the Rogue's skill in thievery, the greater chance of stealing an object from its targets inventory.
* Feedback is provided for successful attempts, unsuccessful, and if the target has nothing to steal.

This has been modified so that it works with 3.6.x, and I have also made the feedback more random as well as adding feedback for monsters with no inventory.

Limitations: this only works on hostile targets. I'd like to work on this some more, make it so that a rogue could attempt to pickpocket a peaceful target also, and would turn hostile if they discover they've been pickpocketed.

I'd also like dexterity and/or luck to play a factor in a successful attempt too, but that plus the above will be for a later day.