Partial fix for 'impossible: Can't place branch!' panic messages. As far as I know, this only occurs in Gehennom on the level where the branch stairs to Vlad's tower is created. The error will still occur, but the player can continue on and use the branch stairs - it will be surrounded by lava and may require a bit of work to access it, but the stairs can still be used.
The partial 'fix' is that the game will no longer lock up and go into an infinite loop writing the same error over and over again to the paniclog file. This obviously needs to be addressed. What I think is happening - the level is made, and the random Gehennom 'hellfill' levels have a random lava river running top to bottom, its length and width random every time. The section of code that determines where the branch stairs are placed is unaware of the random lava placement, hence the panic error. I actually like it that the stairs could sometimes be surrounded by lava, but this panic error needs to be resolved. More to follow.