New player race: Illithid. Illithid is the formal name for a mind flayer. This ancient race has psychic abilities and are natural spellcasters, but are physically weak compared to other races. Here are the details...
* Illithids can only be of chaotic alignment, and can be both male or female
* Illithids can play as either a healer, priest or wizard
* Intelligence and wisdom can go as high as 22, but their strength, constitution and charisma are capped at 13/14/16 respectively
* Power generation per level gained is slightly higher than that of the elf player race
* Illithids have a natural psionic attack that can confuse and kill, and is tied to spell power use
* Illithids start out with psychic resistance (new resistance), infravison and telepathy. At experience level 12 they can fly much like their more wild brethren

There were a couple different ways I could have gone about implementing the Illithids psionic attack, and I wound up settling on it being an actual spell so that its use would consume power. So there's an actual spellbook of 'psionic wave' but there are a few cavets to make it unique to Illithids only. Only the Illithid race starts out with this spellbook in its inventory (and thereby automatically knowing it), and only an Illithid can successfully read it. The spellbooks random spawn probability is zero. This 'spell' will always have 100% retention, and it will always have a 0% chance of failure to 'cast'. Attack characteristics - a flat 2d6 physical damage that currently does not scale, and the attack has a 25% chance of confusing its target. Mindless monsters or monsters resistant to psychic attacks are immune to this attack.Power consumed is 5 per attack (treated as a level 1 attack spell).

One thing I want to implement but haven't figured out yet, is the mind flayers tentacle attack. That will make it in eventually, but it will not be nearly as powerful as the current mind flayers multiple tentacle onslaught. I have already added logic that would automatically prevent a player from using a tentcle attack against a monster it shouldn't (read: footrice).

More to do as this is just the initial commit, but hopefully this will prove to be a fun race to play without being unbalanced or too powerful.