Convict patch. This commit adds a new role to the game - the Convict. A somewhat popular role that can be found in other variants like UnNetHack, dNetHack, DynaHack and recently SpliceHack. The version for EvilHack has been adjusted slightly. The details:
* Convicts can be played as a human, dwarf, gnome, orc or hobbit
* Their alignment is always chaotic, regardless of racial background
* Convicts start the game with a negative alignment, a heavy iron ball chained to their leg, and a cursed striped shirt (worn). The shirt prevents entry into any shop while worn
* Convicts start as almost hungry, but can be in a state of hunger for a longer period of time
* Convicts start with intrinsic sickness resistance, and gain intrinsic poison resistance and searching at experience levels 7 and 20 respectively
* Their first artifact gift is Luck Blade, a chaotic broadsword (short sword in other variants) that acts as a luckstone
* The Iron Ball of Liberation is the convict's quest artifact - a heavy iron ball that grants magic resistance, searching, warning and stealth when carried. When invoked, it allows phasing temporarily just like a Xorn, but it attaches itself to you upon every use
* Their starting pet is a sewer rat, and convict's can tame other rats by #chatting them

A few new monsters were introduced as well that are typically only found in the Convict's quest. The enormous rat and rodent of unusual size I had already added a couple months ago, but they already fit in nicely with this patch. Its player monster counterpart was also created along with this patch.

The Convict role was designed to be one of the most challenging roles to play. Couple that with this variant, this should prove to be a considerable feat worthy of bragging rights should anyone ascend a convict.

Expect a couple more minor tweaks, but overall the inital commit is solid. Only issue I am currently aware of, is that you can't wish for the convict's quest artifact. Another weird to-do in src/objnam.c