New monster: Sea Dragon. The sea dragon is of a special sort - its body is made almost entirely out of water, much like a water elemental. Its breath attack is a powerful blast of water that slams into its target, dealing physical damage and has the potential to rust unprotected weapons and armor in open inventory. As all dragons in EvilHack can engulf, the sea dragon can as well, but take heed - its engulf attack is the same as a water elementals engulf attack. If you do not have a means to breathe underwater (or forego breathing altogether), you will suffocate unless you can kill the dragon first or escape from inside of it. Its physical attacks have a chance to do rust damage to weapons/armor, and the sea dragon has a passive AD_RUST attack too.
Now for some of the positives as far as the player is concerned: sea dragons prefer pools of water, and will tend to stay in those if they are available. They are the only dragons that cannot fly, however they can swim and are amphibious. They can be ridden as a mount with the proper gear. The scales that sea dragons drop confer magical breathing, and its secondary intrinsic allows the wearer to swim.

I imagine this creature will be tweaked a bit more in the near future. I don't think there's anything quite like this dragon in any variant, or if there is I'm unaware of it. Hopefully this turns out to be a fun and challenging dragon, whose scales are worth wearing.