Fix: how helmets are supposed to block attacks to your head. Funny story... I bring over code from GruntHack for the zombie bite attack that can make the player ill and turn into a zombie. I inadvertently put the block of code that determines whether the bite attack makes you diseased or not *above* the block of code that gives head gear a chance to block attacks to the players head. This means all zombie bite attacks were being processed first (80% chance of a bite attack making you diseased) before it went to the helmet blocking bit. GruntHack is the exact same way, I copied it over in the same order and didn't realize that wearing a helmet was hardly blocking any zombie attacks at all. In fact it never did - if the 1 in 5 chance of not being diseased happened, *then* it could progress down and have the 7 in 8 chance of the players helmet block an attack. Wow.
So, this commit corrects that oversight. And the head gear material can determine how well it will block an attack to the head (4 in 5 chance of success for any hard material, 1 in 2 chance of success for any other material). This materials change affects blocking any head attack, not just from zombies bite attack.

Also in this commit: better feedback when the player's brain is eaten by a zombie, and I've increased the kobold zombie difficulty level by one. This has the following effect - the player won't see any zombies on dungeon level 1 unless they park there and level up several experience levels.

This commit will have a significant impact on the early game - a break from zombies on the first level, and wearing something on your head actually protects you to a certain degree.