New object: crystal chest. The crystal chest is a very rare and special type of magical container. It is made of gemstone, is white in color (imagine it's made of dilithium), and weighs slightly less than your average chest. This is what is special about it:
* Crystal chests cannot be locked nor unlocked by any physical means
* Trying to physically #force open a crystal chest is also not possible
* Crystal chests do not generate as trapped
* Crystal chests can only be locked or unlocked by magical means
* Crystal chests are immune to most other forms of magic

What this means - this chest is immune to polymorph and probing. It also
cannot be unlocked by intelligent monsters with keys or other lock
opening tools. There currently isn't any logic for intelligent monsters
to want to use wands of opening, nor can spellcasting monsters cast
anything remotely like the spell 'knock' (for now). This basically makes
a crystal chest the ideal container to stash all of the items you want
to keep and not carry around.

This could use some more work - currently you can't lock/unlock this
chest with The Master Key of Thievery as it's a physical function, and I
should add a special condition for it. I may want to make it immune to
other forms of magic as well.