Silencing compiler warnings: unused variables. The variant authors of SplceHack and xNetHack use a series of warning flags in their debugging hints file to help them catch any issues that could cause problems down the road. I used to think I had a clean build until using these (see sys/unix/hints/linux for the additional CFLAGS).
There's a fair amount to clean up. A small bit exists in vanilla, the
rest is on me. This commit cleans up almost all of the 'unused variable'
warnings during compile. I left one alone (src/files.c) because it could
be used by someone else but isn't because of an #indef statement.