Naming/wishing for artifacts tweaks. Wishing - wishing for an artifact will now always have its owner with it, no more maybe getting away with having two or three artifacts before you have to deal with its owner.
Naming - you can still name an elven dagger 'Sting' or an elven
broadsword 'Orcrist' and you'll still get it. But an envoy representing
elvenkind will show up, and they're not too impressed. Nothing as
difficult as a player monster or a quest nemesis, but you'll have to
fight for what you just obtained.