New dungeon features: shallow water and sewage, new monsters: giant cockroach, giant leech, giant crocodile, new mines' end map: The Sewers of Waterdeep. Introducing two new dungeon features to the game - shallow water and sewage. Shallow water comes from the terrain patch by L - I used it as a template as well as some of the changes for this found in NetHack Fourk, and then modified a bit further by me. Shallow water has the following properties:
* Shallow water can be walked across without sinking as long as you (or
a monster) are larger than MZ_TINY. Tiny creatures will not attempt to
cross shallow water via walking
* You can dip objects into shallow water just like you would a pool or
moat
* Walking in shallow water can rust susceptible footwear if not
fooproofed
* Traps can form underneath shallow water (beware)
* Shallow water can sometimes form around fountains and sinks. Sometimes
piranha will appear along with the shallow water formed

Sewage is an entirely new custom feature - it shares many of the same
properties as shallow water, with the following additions:

* You can dip objects into sewage just as you can shallow water, but I
wouldn't advise it
* You can also attempt to drink sewage. Don't advise doing this either
* Walking through raw sewage is a lot of work as it is much thicker than
regular water - the player is slowed as a result until they stop walking
in it
* Some creatures prefer pools of raw sewage and will migrate to it and
stay in it if possible

Both shallow water and sewage have been added to a variety of maps, such
as the various quest maps that feature pools, certain custom maps in
Gehennom, and the rivers that run through the Gnomish Mines now have a
small chance of being sewage now instead of lava (same rivers feature
shallow water with regular pools).

I've created a new custom mines' end map to showcase sewage along with
three new monsters - giant cockroaches, giant leeches, and giant
crocodiles. The first two do not spawn at random and are currently only
found with this new mines' end map and a couple other areas with sewage.
Giant leeches prefer to be in raw sewage and will hide under the surface
to attack and stick to the hero. Giant cockroaches can slip under
objects and hide that way. Both have a bite much the same as a rabid rat
(poisoned bite that can drain constitution). Giant crocodiles can spawn
randomly - they are just a larger and slightly more powerful version of
a regular crocodile, but I may revisit this particular creature at a later
date.

The new mines' end map is called 'The Sewers of Waterdeep' - standing
sewage is everywhere, and the level is infested with rats of various
types, cockroaches, leeches, and a few crocodiles for good measure. The
hero will need to traverse the sewer pipes to find the luckstone and any
other treasure that may have made its way down to the sewers. This can
be a challenging map, as walking in sewage will slow the player down,
and there aren't many dry patches of floor available. Players will need
to be a bit creative and plan ahead to survive the sewers of Waterdeep.

I hope players enjoy these additions, was a fair amount of fun to
create.