New unique monster: The Rat King. Tangles gave this idea after he had played the new mines' end sewer level and had found the throne hidden there. Enter The Rat King:
* The Rat King is flagged MH_WERE but will never change form, otherwise
it shares all of the same abilities, strengths, and weaknesses as any
other werecreature
* The Rat King is poison and sleep resistant, is slightly magic
resistant, and has two forms of attack (one weapon attack, and a bite
attack that can cause lycanthropy)
* The Rat King spawns wielding a poisoned scimitar, and wearing high
boots and a cloak of displacement

This unique monster isn't anything too crazy - its level is appropriate
for where it's found. But it does add a bit more flavor to the Sewers of
Waterdeep and makes being able to use the throne on that level a bit
more of a challenge.