Fix: chest containing the wand of wishing on castle-3.lev would sometimes spawn as trapped. Oversight on my part when importing the castle level into EvilHack - none of the chests that the wand of wishing spawns in is supposed to be trapped.
Took this a step further - on all castle levels, there's a 20% chance
that chest the castle wand of wishing spawns in is a crystal chest
instead of a regular chest. The upside to this is that no monster has
the ability to get into this type of chest, plus it's immune to
cancellation and polymorph magic. The downside is that if the player
doesn't have a wand of opening or the spell knock learned, they're not
getting into it either. So come prepared.