Fix: another artifact wishing revamp. This is labelled as a fix because when doing the revamp, I noticed a bug that could cause a crash under the right conditions.
The revamp part - have heard from various players and have also noticed
myself, that if the player were strong enough to handle high-level
player monsters, they could wish for as many artifacts as they had
wishes to burn, and could get every artifact along with all the
ascension-level gear that the owner of the artifact comes with. This was
never the intent. In this commit, we're reverting artifact wish chances
back to the way vanilla has it (see
https://nethackwiki.com/wiki/Wish#Artifact_wishing for the odds based on
how many artifacts have been generated). Getting the owner of said
artifact is still a possibility if the player successfully wishes for
an artifact, and is also tied to how many artifacts have already been
generated - about a 50% chance of getting the owner if it's the first
artifact generated, and the odds scale up from there. This is slightly
more forgiving on getting the owner, but the player can no longer wish
for as many artifacts as they want.

The fix part - the envoy for Vlad if you wished for Lifestealer was a
type of vampire. If that type had already been genocided, the game would
crash on a successful Lifestealer wish. This commit changes the envoy to
a monster that cannot be genocided, and is also able to wield
Lifestealer.