Monks can break a monster's wielded weapon under certain conditions. This was a YANI I mentioned in IRC a long time ago, forgot about it, then was shown it again recently.
Monks can break a weapon wielded by its target under certain conditions.
Those conditions are:
* the Monk is of grand master skill in martial arts, is fighting bare
handed, and is not wearing body armor nor a shield
* the monsters weapon cannot fall under is_flimsy()
* if the monster's wielded weapon is damaged, it increases the odds of
its weapon breaking

The odds are the same as if wielding a two-handed weapon or if playing
as a Samurai and wielding a katana (vanilla NetHack). I also added a bit
of extra verbage to when a weapon is broken based on its material.