more appropriate non-artifact sacrifice gifts
Ideally this will fix the following issues:
 - Giants and centaurs shouldn't be able to get armor they can't wear (should have been already fixed, cleaner now though)
 - Elves and orcs shouldn't get items whose material they hate (mainly elves with iron)
 - (minor issue) racial gear should only be given to that race (so a dwarf won't get an elven dagger), and if possible convert standard gear into racial (dagger -> elven dagger if you're an elf, spear -> orcish spear for orcs, etc.)

Does NOT touch the spellbook giving code at all, except maybe some whitespace. Tested as best I could, no crashes and it seems to work as intended.

Rarely, elf PCs with awful weapon skills (mainly rangers, seems like) can get REALLY unlucky rolls on weapons (and not get any). This is because the code just clears racial gear that doesn't match the race (so if they get an orcish one it won't convert to elven, just reroll completely), and if you happen to not land on a weapon you can use then you're out of luck. To compensate, the rolls has been upped to 1000 (and a debug message added if it bypasses that). This should only really be a problem with maybe rangers and rogues, anybody who doesn't get too many weapon skills (of weapons possible to get gifted, spear through katana).