Partial fix: active extrinsic from offhand weapon but not twoweaponing. Before this commit, players could twoweapon artifact weapons or weapons with various object properties (searching, warning, or ESP) and still enjoy the extrinsic benefits of the offhand weapon after switching back to their primary weapon and no longer twoweaponing. Decided to have another stan at correcting this, and the current state is this:
* the player no longer gets the extrinsic from the offhand weapon when
not in twoweapon mode
* when the player enters twoweaponing mode, the offhand artifact or
weapon with object properties won't activate its extrinsic properties
unless the player saves and reloads the game

The changes in this commit are virtually identical to how this is
handled in GruntHack, and it behaves as it should without having to
save/reload a game with that variant. At this point I'm kind of stumped
and could use another set of eyes on it.

There's also a bit of formatting clean-up in this commit as well.